requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.5
httpx>=0.27
pytest>=8.0
anthropic>=0.40
streamlit>=1.30
pyyaml>=6.0
huggingface_hub>=0.24
